Added StrokeAsset properties to TonalArtMapGeneratorWindow, Fixed weird First/Last tones in TonalArtMapGenerator
TonalArtMapGeneratorWindow selectively displays a StrokeAsset's data if available. StrokeAsset has it's own editor to allow for multi usage in editor window + scriptable inspector. Additionally, fixed a logic error causing the first and second tones to always be white - and updated TonalArtMapGenerator to be able to force - or not - the first tone to be full white and the last tone to be full black.