Add Tonal Art Map wrapper settings to TonalArtMapGeneratorWindow
Display TonalArtMapAsset in the tool editor window, with wrapped properties to ensure proper handling of the asset settings. Swapping the number of the expected tones no longer results in the renderer feature considering less or more tones to be present. TODO: Display a warning telling the user that the number of underlying tones in the asset is different from the one in ExpectedTones, asking for a rebake of the textures.