Finalize TonalArtMapGeneratorWindow, Create SketchRendererUI UIToolkit Library.
Finalized the first iteration of the visual layout for the TonalArtMapGeneratorWindow and all supported functionality, such as generating TAMS, a live preview of the texture, and automatically creating and assigning TAMs to the sketch renderer.

Additionally, this also includes the core of SketchRendererUI UIToolkit elements library, which provides ready-made implementations and structures for inspector elements to be used by future editor windows/inspectors.

Fixed a bug in TonalArtMapAsset refactor causing sketch renderer to consider the number of textures as the number of packed elements in the entire asset.