Finish all previous feature ports, Editor Tool Windows, Sketch Manager (#22)
* Created system for loading SketchResourceAssets from Package files. Updated RendererFeatures to load asset dependencies when added through setup wizard. Fixed AccentedOutlines clearing the baked UV texture. Ported TonalArtMapGenerator, updated to exist within Package. Created system for Package EditorWindows, added TonalArtMapEditorWindow (WIP).

* Added StrokeAsset properties to TonalArtMapGeneratorWindow, Fixed weird First/Last tones in TonalArtMapGenerator

TonalArtMapGeneratorWindow selectively displays a StrokeAsset's data if available. StrokeAsset has it's own editor to allow for multi usage in editor window + scriptable inspector. Additionally, fixed a logic error causing the first and second tones to always be white - and updated TonalArtMapGenerator to be able to force - or not - the first tone to be full white and the last tone to be full black.

* Add Tonal Art Map wrapper settings to TonalArtMapGeneratorWindow

Display TonalArtMapAsset in the tool editor window, with wrapped properties to ensure proper handling of the asset settings. Swapping the number of the expected tones no longer results in the renderer feature considering less or more tones to be present. TODO: Display a warning telling the user that the number of underlying tones in the asset is different from the one in ExpectedTones, asking for a rebake of the textures.

* Fixed Tool Window Initialization on Editor Opening. Created CustomEditors for StrokeAsset inheritors. Additional features for TonalArtMapAsset in editor window.

Fixed an early initialization issue causing TextureToolWizard to not find existing tools in active layout due to editor not being initialized.

Created additional editors for Hatching, Zigzag and Feathering stroke assets to display their additional features.

TonalArtMapAssets can now be created and set to current renderer context dynamically through the TonalArtMapGeneratorWindow. Also fixed an issue causing NumberOfTones to not reflect the reality of the asset after reopening the project.

* Finalize TonalArtMapGeneratorWindow, Create SketchRendererUI UIToolkit Library.

Finalized the first iteration of the visual layout for the TonalArtMapGeneratorWindow and all supported functionality, such as generating TAMS, a live preview of the texture, and automatically creating and assigning TAMs to the sketch renderer.

Additionally, this also includes the core of SketchRendererUI UIToolkit elements library, which provides ready-made implementations and structures for inspector elements to be used by future editor windows/inspectors.

Fixed a bug in TonalArtMapAsset refactor causing sketch renderer to consider the number of textures as the number of packed elements in the entire asset.

* Merge from feature/material-editor-window

* Ported MaterialGenerator and Created MaterialGeneratorEditorWindow

* Create New MaterialData and assign generated textures to current renderer context from window.

Added dynamic check to display a create and assign MaterialDataAsset to the window when no value is assigned.

Allow generated textures to automatically be added to current SketchRendererContext's MaterialSurface render pass data.

Some additional fixes to missing generating behaviour in editor window.

* Prevent TextureGenerator updates from one window from repainting the other. Added a create stroke asset wizard to editor window.

* Initialize Created Scriptables from Default Instances

* Port features to rendergraph, fix stroke composition, fix baked distortion becoming invalid. (#21)

* Port remaining features to rendergraph support, fix stroke composition, fix baked distortion buffers becoming invalid.

Ported all edge detection variant passes and other renderer features which were made using native passes to rendergraph wherever possible, to allow for automatic pass combination.

Fixed dynamic scaling causing baked distortion buffers from being cleared and not rebaked. The buffers are now created at a static resolution and are checked if rebaking is necessary dynamically at execution.

Also made the stroke blending with luminance and material not break completely when using non-dark colors (by actually doing real blending).

* Add Baked Distortion Scale to SmoothOutline volume.

Implement connection between a volume override and the pass data in AccentedOutlinePassData.