Prevent editing of inmutable assets
Added new checks to detected if the current asset being used by a tool is inmutable (default), and prevent modifying it since it would just get overwritten anyways.

Also updated SketchRendererManagerSettings to be created dynamically when adding the package, since the inmutable settings went agaisnt the idea of saving configuration.